Move runtime converters to the io.vertx.codegen.protobuf package
