+++ vertx-codegen-protobuf/src/main/java/io/vertx/codegen/protobuf/converters/JsonObjectProtoConverter.java
@@ -1,4 +1,4 @@
-package io.vertx.core.proto;
+package io.vertx.codegen.protobuf.converters;
 
 import com.google.protobuf.ByteString;
 import com.google.protobuf.CodedInputStream;
